+++ lib/syncmsk.f90
@@ -262,7 +262,7 @@ subroutine syncmsk(cdat,npts,jpk,ipk,idf,rmax,snr,metric,decoded)
      call vit213(e1,nb1,mettab,d8,metric)
      call system_clock(count1,clkfreq)
      tvit=tvit + (count1-count0)/float(clkfreq)
-     ihash=nhash(c_loc(d8),9,146)
+     ihash=nhash(c_loc(d8),int8(9),146)
      ihash=2*iand(ihash,32767)
      decoded='                      '
      if(d8(10).eq.i1hash(2) .and. d8(11).eq.i1hash(1)) then
